+++ changelog.md
@@ -1,11 +1,20 @@
 <!--
   Created at: 08/05/2021 14:16:41 Thursday
-  Modified at: 08/07/2021 07:58:43 AM Saturday
+  Modified at: 08/08/2021 06:06:40 PM Sunday
 -->
 
 # Changelog
 
 A verbose logging of every change.
+
+## Version 0.2.3 (08/08/2021 18:06:39 Sunday)
+
+- Fixed keywords parsing
+- Added keywords to tests
+- Changed thumbnails length test to be more than 1 (not 0)
+- Added test for `publishDate` and `uploadDate`
+
+---
 
 ## Version 0.2.2 (08/07/2021 07:58:06 Saturday)